requirements.txt
lxml==5.3.0
python-docx==1.1.2
typing_extensions==4.12.2
wrapt==1.16.0
